requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
pyyaml
httpx
httpx
